vmwareapi oslo.vmware library integration
These are the minimal changes necessary to integrate the vmwareapi
driver with the oslo.vmware code. Much of the change is mechanical
switching to call the library code, and removing obseleted code in the
driver where appropriate.

Effort was made to minimize the amount of actual driver code changes in
this round by keeping a small amount of shim code. Existing unit tests
were modified as appropriate. Additional patches will be introduced
shortly to include additional conversions and tests.

Summary of changes:

- use common/vmware/exceptions.py - there are a linger bit of code (most
  for defining nova-specific exceptions that are still retained in
  error_util.py)
- use oslo.vmware vim.py for the Vim implemenation
- use oslo.vmware api.py for api invocation/session
- updated existing unit tests
- nova retains its own version of get_{soap/wsdl}_url which using conf
  settings for default and overrides.

Change-Id: Idab7778d4c7e5e7099d842debbcbc009cf9127b6