VMware: Copy image handler
The VMware driver should take advantage of the Glance VMware
datastore storage backend in order to avoid a download of the
image from the Glance store to the VMware datastore.
This patch is the implementation of a new image handler using
the location URL provided by the V2 API of Glance to directly
copy the file on the VMware datastore in order to spawn the instance.
If this copy fails, the download image handler will be used.

The patch partially implements bp: image-multiple-location
and bp: nova-image-zero-copy

The next step would be to use directly the image without any copy.

Change-Id: I6c4231c70ef4e6a95817fdb3baa2725837f2003e

Conflicts:

	nova/tests/virt/vmwareapi/test_driver_api.py
	nova/virt/vmwareapi/imagehandler/download.py
	nova/virt/vmwareapi/vmops.py