vmwareapi: enable spawn from ova image
The vmware driver has not been honoring the container_format of "ovf".
The change narrowly interprets any image with this container_format
value to be a single-file OVA image, and will attempt to stream the
first disk in the OVA as the root disk image.

The disk is guaranteed to be in streamOptimized format, which is
currently the only format we can deploy on regular and vSAN datastores.

Closes-Bug: #1286405
Change-Id: Iddb7cca8af755f6cc0eb6861bce6cb29a4c18b50