vmwareapi: snapshot as stream-optimized image
This changes the behavior of the nova snapshot operation to upload
the snapshot as a stream-optimized vmdk disk.
It does so by takin a temporary VM snapshot of the instance VM, create a
VM clone out of said snapshot, and using ExportVM of the clone to obtain
a HTTP stream of the stream-optimized disk.

Implements: blueprint vmware-vsan-support
Change-Id: Ic4f4fb2411b2fddd98f95c24b815176dff3acb26