VMware: Clone snapshots to the Glance datastore
Glance contains a new backend store since Icehouse: the VMware store.
The Nova VMware driver should take advantage of it for snapshots:
instead of uploading the bits through the network, we move them
to the Glance datastore folder and register a new location for the
image. This makes the snapshot operation much faster.

This optimization is built on top of the image handler framework
using the push_image() api.

partially implements bp vmware-clone-image-handler

Change-Id: I7fdbed84749cfa9d9a440ba591a24bbac522d7ff